Bug 15091: Batch Checkout: Set titles above result tables
This patch is to clarify the results that are displayed while performing
a batch checkout.

It is not the full solution for what is described in the video in
comment #2, but it tells the user what the information in the table
is about.

A second patch is planned to address the case where the issues are confirmed
but not can be fulfilled because the user's fines reach the threshold while
checking out the items.

Signed-off-by: Nicole Engard <nengard@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>