Bug 15256: Fix items table display on the detail page
Bug 14966 fixed a bug but introduced another one: With
IndependentBranches set to 'prevent', if the librarian (without the
superlibrarian permission) can edit items but not all, the table is
broken.  DataTables raises the following warning: "DataTables warning:
table id=DataTables_Table_1 - Requested unknown parameter '8' for row0."
And does not display correctly.

Test plan:
1/ Create a biblio record with 1+ items located in different places.
2/ Create/use a patron with the permission to edit items
3/ Set the IndependentBranches pref to 'prevent'
4/ With this patron logged in, go on the record detail page.
Without this patch, the display is broken (only 1 edit link and 1 td
missing, DataTables raises a warning).
With this patch, the display is correct.

Signed-off-by: Frédéric Demians <f.demians@tamil.fr>
  See the warning, and the broken table. Solved by this patch.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>