Bug 15313 - Use Bootstrap modal for z39.50 search MARC and Card preview
In order to reduce the number of dependencies for common interactions we
should attempt to use Bootstrap modals in place of Greybox ones.

This patch modifies the cataloging Z39.50 search results page so that
"MARC" and "Card" previews are displayed in Bootstrap modals.

Because the modal is triggered from a pop-up window, the modal has been
reformated to exclude the footer and reduce the size of the modal
header.

To test:

1. Apply the patch and navigate to Cataloging.
2. Click the "New from Z39.50/SRU button to open the search
   window.
3. Perform any search which will return results.
4. Test the visible "MARC" and "Card" links for multiple results and
   confirm that the correct data is shown each time.
5. Click anywhere in the results table to show the hidden menu and test
   the "MARC" and "Card" links it displays.

Corrected: The original patch didn't address the hidden menu which
appears if you click in any table cell.

Followed test plan, works as expected (including click in table cell).
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>