Bug 15318 - Use Bootstrap modal for MARC and Card preview when ordering an external source
In order to reduce the number of dependencies for common interactions we
should attempt to use Bootstrap modals in place of Greybox ones.

This patch modifies the page in Acquisitions for ordering from an
external source so that the "MARC" and "Card" previews display in
Bootstrap modals.

This patch also fixes a bug where the in-table popup menu points to the
wrong record (fixed in Cataloging by Bug 6367).

To test:

1. Apply the patch and navigate to Acquisitions.
2. Locate an exsiting basket or create a new one.
3. Add to the basket using the option to order from an external source..
4. Perform a search which will return multiple results.
5. Test the "MARC" and "Card" links for multiple results and confirm
   that the correct data is shown each time.
6. Click anywhere in the results table to show the hidden menu and test
   the "MARC" and "Card" links it displays, confirming that the data
   looks correct and corresponds correctly to the title you selected.

Followed test plan, works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>