Bug 15317 - Use Bootstrap modal for MARC and Card preview when ordering from staged files
In order to reduce the number of dependencies for common interactions we
should attempt to use Bootstrap modals in place of Greybox ones.

This patch modifies the page in Acquisitions for ordering from a staged
file so that the "MARC" and "Card" previews display in Bootstrap modals.

To test:

1. Apply the patch and navigate to Acquisitions.
2. Locate an exsiting basket or create a new one.
3. Add to the basket using the option to order from a staged file.
4. Click "Add orders" from any staged file.
5. Test the "MARC" and "Card" links for multiple results and confirm
   that the correct data is shown each time.

Followed test plan, works as expected
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>