Bug 15375 - Translatability: Fix issues on OPAC page 'Placing a hold'
Fix several issues with odd translations on opac-reserve.tt due to
string splitting by tags.
In addition, streamline messages to make them similar to the same
messages on the OPAC summary page (Bug 15374)

To test:
- Apply patch
- Log in to OPAC
  - with blocked /debarred user (with and without comment and date)
  - with user who has to much fines
  - with user who has his card marked as lost
  - with user with expired account
  - with user with uncertain address information
- Try to put holds
- Verify that message about why holds are not possible is correct
  and displays nicely and that links in messages work
- Carefully examine code
- Search for regressions

(Amended: stray i removed, test plan updated, removed account that
          is about to expire)

Signed-off-by: Aleisha <aleishaamohia@hotmail.com>

I am going to sign off on this because it works as it should and I think the problems I had in comment 8 are a separate bug.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>