Bug 15380: Use Koha::Authority::Type[s] in admin script
This path rewrites the admin/authtypes.pl script to use the new modules
instead of executing the SQL queries from the script.

Test plan:
1/ Create 1+ authority types
2/ Update one
3/ Delete it
4/ Create an authority record using this authority type.
5/ Try to delete the authority type
Here there is a change in the behavior.
Prior to this patch, the user was able to delete the authority type, now
it's not.
I think it was a bug, feel free to give your opinion if you disagree on
this change.

Signed-off-by: Marc Véron <veron@veron.ch>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com