Bug 15469: Fix authority header search
Introduced by bug 15381.

The logs contain plenty of "detail.pl: No method value!" and the
dropwdown list is not correctly populated (no value for options)

Test plan:
Go to the authority module and have a look at the auth type codes
dropdown list.
Without this patch, the values of the options are empty.
With this patch, the values are correctly filled.

Signed-off-by: Frédéric Demians <f.demians@tamil.fr>
  I confirm this patch silence noise in log file.

Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com